Refactor: 내 프로필, 유저 프로필 컴포넌트 병합 1차 리팩토링
