+++ src/api/profile.js
@@ -1,6 +1,6 @@
 const PROFILE_URL = "https://api.mandarin.weniv.co.kr/user/myinfo"
 
-const profileAPI = async (token) => {
+export const profileAPI = async (token) => {
   try {
     const response = await fetch(PROFILE_URL, {
       method: 'GET',
@@ -16,6 +16,21 @@ const profileAPI = async (token) => {
   }
 };
 
-export default profileAPI;
-// test 계정 정보 : suritest@test.com / suritest
-// test 계정 정보 : wonbeom@weniv.co.kr / 123123
+export const accountProfileAPI = async (accountname, token) => {
+  try {
+    const ACCOUNTPROFILE_URL = `https://api.mandarin.weniv.co.kr/profile/${accountname}`
+    const response = await fetch(ACCOUNTPROFILE_URL, {
+      method: 'GET',
+      headers: {
+        "Authorization": `Bearer ${token}`,
+        "Content-type": "application/json"
+      }
+    })
+
+    const data = await response.json();
+    console.log(data);
+    return data;
+  } catch (error) {
+    console.error('Account API 에러가 발생했습니다', error);
+  }
+};
